+++ src/Actions/EnableTableChangeTracking.php
@@ -5,41 +5,31 @@ namespace Patabugen\MssqlChanges\Actions;
 use Illuminate\Console\Command;
 use Illuminate\Database\QueryException;
 use Illuminate\Support\Str;
+use Patabugen\MssqlChanges\Table;
 
 class EnableTableChangeTracking extends BaseAction
 {
-    public string $commandSignature = 'mssql:enable-table-change-tracking {table}';
-
     private array $messages = [
         1 => 'Change Tracking enabled for table %s',
         2 => 'Change Tracking is already enabled for table %s',
     ];
 
-    public function handle(string $rawTableName): string
+    public function handle(Table $table): string
     {
-        // Some rudimentary sanitising/escaping.
-        $tableName = Str::of($rawTableName)->ascii()->wrap('[', ']');
-
-        throw_unless(
-            $this->connection()->getSchemaBuilder()->hasTable($rawTableName),
-            'Table '.$tableName.' does not exist'
-            .' in database '.$this->connection()->getDatabaseName()
-            .' at '.$this->connection()->getConfig('host')
-        );
-
+        $tableName = Str::of($table->name)->ascii()->wrap('[', ']');
         try {
             $this->connection()->unprepared(
                 'ALTER TABLE '.$tableName.' ENABLE CHANGE_TRACKING WITH (TRACK_COLUMNS_UPDATED = ON);',
             );
         } catch (QueryException $e) {
             if (Str::of($e->getMessage())->contains('Change tracking is already enabled for table')) {
-                return $this->return($this->messages[2], $tableName);
+                return $this->return($this->messages[2], $table->name);
             } else {
                 throw $e;
             }
         }
 
-        return $this->return($this->messages[1], $tableName);
+        return $this->return($this->messages[1], $table->name);
     }
 
     private function return(string $messageTemplate, $databaseName): string
@@ -50,10 +40,32 @@ class EnableTableChangeTracking extends BaseAction
         return $message;
     }
 
+    public function runAllTables()
+    {
+        $tables = ListTables::run();
+
+        return $tables->map(function (Table $table) {
+            return $this->handle($table);
+        });
+    }
+
     public function asCommand(Command $command): void
     {
-        $command->info($this->handle(
-            $command->argument('table')
-        ));
+        $tables = [];
+        if ($command->option('all')) {
+            $tables = ListTables::run();
+        } else {
+            $tables = collect([
+                Table::create($command->argument('table')),
+            ]);
+        }
+        $tables->each(function (Table $table) use ($command) {
+            if (empty($table->primaryKeyName)) {
+                $command->warn($table->name.' cannot have Change Tracking enabled because it does not have a Primary Key');
+
+                return;
+            }ray($table);
+            $command->info($this->handle($table));
+        });
     }
 }