+++ src/Actions/ListTables.php
@@ -8,9 +8,9 @@ use Patabugen\MssqlChanges\Table;
 
 class ListTables extends BaseAction
 {
-    public string $commandSignature = 'mssql:list-tables';
-
     public array $tableFilter = [];
+
+    public bool $onlyTrackingEnabled = false;
 
     /**
      * Returns a collection of Table objects for any table which has Change Tracking
@@ -21,15 +21,22 @@ class ListTables extends BaseAction
     public function handle(): Collection
     {
         $query = $this->connection()
-            ->table('sys.change_tracking_tables')
-            ->select('sys.tables.*')
+            ->table('sys.tables')
+            ->select(['sys.tables.*', 'sys.change_tracking_tables.is_track_columns_updated_on'])
             // Link the change-tracking system info to the tables list
-            ->join('sys.tables', 'sys.change_tracking_tables.object_id', 'sys.tables.object_id')
+            ->leftJoin('sys.change_tracking_tables', 'sys.tables.object_id', 'sys.change_tracking_tables.object_id')
             ->orderBy('sys.tables.name')
+            ->when($this->onlyTrackingEnabled, fn ($query) => $query->where('sys.change_tracking_tables.is_track_columns_updated_on', '=', '1'))
             ->when(! empty($this->tableFilter), fn ($query) => $query->whereIn('sys.tables.name', $this->tableFilter));
 
         return $query->get()->mapWithKeys(function ($item) {
-            $primaryKey = $this->connection()->select('EXEC sp_pkeys ?', [$item->name])[0]->COLUMN_NAME;
+            $primaryKey = collect($this->connection()->select('EXEC sp_pkeys ?', [$item->name]));
+
+            $primaryKey = ($primaryKey->isEmpty())
+                ? ''
+                : $primaryKey->pluck('COLUMN_NAME')->join(',');
+
+            $changeTrackingEnabled = $item->is_track_columns_updated_on == '1';
 
             $columns = $this->connection()
                 ->table('INFORMATION_SCHEMA.COLUMNS')
@@ -41,7 +48,7 @@ class ListTables extends BaseAction
                 $item->name => new Table(
                     connection: $this->connection(),
                     name: $item->name,
-                    columnTrackingEnabled: true,
+                    columnTrackingEnabled: $changeTrackingEnabled,
                     primaryKeyName: $primaryKey,
                     columns: $columns
                 ),
@@ -56,6 +63,13 @@ class ListTables extends BaseAction
         return $this;
     }
 
+    public function onlyWithTracking($onlyTrackingEnabled = true)
+    {
+        $this->onlyTrackingEnabled = $onlyTrackingEnabled;
+
+        return $this;
+    }
+
     public function asCommand(Command $command): void
     {
         $tables = $this->handle();
@@ -66,6 +80,8 @@ class ListTables extends BaseAction
                 return $table->toArray();
             })->toArray(),
         );
-        $command->info(count($tables).' tables have change tracking enabled');
+        $stats = $tables->countBy('columnTrackingEnabled');
+        $command->info($stats[0].' tables have change tracking disabled');
+        $command->info($stats[1].' tables have change tracking enabled');
     }
 }